chore(xod-cli): move tests to test-func
