+++ packages/xod-cli/package.json
@@ -87,7 +87,7 @@
     "dev": "yarn run build --watch",
     "postpack": "rm -f oclif.manifest.json",
     "prepack": "yarn run build && oclif-dev manifest && oclif-dev readme",
-    "test": "mocha test/**/*.spec.js",
+    "test-func": "BABEL_DISABLE_CACHE=1 mocha --opts test-func/mocha.opts test-func/**/*.spec.js",
     "version": "oclif-dev readme && git add README.md"
   }
 }
