+++ packages/xod-cli/test-func/boards.spec.js
