+++ packages/xod-cli/test-func/compile-upload.spec.js
