+++ packages/xod-cli/test-func/installArch.spec.js
