+++ packages/xod-cli/test-func/mocha.opts
