+++ packages/xod-cli/test-func/publish.spec.js
