+++ packages/xod-cli/test-func/transpile.spec.js
